refactor: merge stateful zone back into common.py
The _stateful_zone.py split left re-export ceremony in common.py and
stranded gp2_gib_for_io — used by rds.py, not stateful-zone code — in
a module named for a different concern. Moving back keeps the two
private helpers adjacent to compute_stateful_zone, drops the
re-exports, and preserves every existing import path.

Also adds section banners and per-step comments inside
compute_stateful_zone so the resource-count breakdown reads linearly,
and tightens the mutually-exclusive disk branches from if/if to
if/elif.
